+++ public/10.js
@@ -1,174 +1,56 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[10],{
 
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=script&lang=js&":
-/*!****************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=script&lang=js& ***!
-  \****************************************************************************************************************************************************************************************************/
-/*! exports provided: default */
+/***/ "./resources/js/components/companyInfo/CompanyInfoData.js":
+/*!****************************************************************!*\
+  !*** ./resources/js/components/companyInfo/CompanyInfoData.js ***!
+  \****************************************************************/
+/*! exports provided: companyInfoData */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-/* harmony default export */ __webpack_exports__["default"] = ({
-  name: "History",
-  props: ["history"],
-  data: function data() {
-    return {
-      isOpen: false
-    };
-  },
-  methods: {
-    isOpenHandler: function isOpenHandler() {
-      this.isOpen = !this.isOpen;
-    }
-  },
-  computed: {
-    slicedHistory: function slicedHistory() {
-      return Object.values(this.history).slice(0, 2);
-    }
-  }
-});
-
-/***/ }),
-
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=template&id=7201f5a7&":
-/*!********************************************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=template&id=7201f5a7& ***!
-  \********************************************************************************************************************************************************************************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "render", function() { return render; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return staticRenderFns; });
-var render = function() {
-  var _vm = this
-  var _h = _vm.$createElement
-  var _c = _vm._self._c || _h
-  return _c("div", [
-    _c(
-      "ul",
-      _vm._l(_vm.isOpen ? _vm.history : _vm.slicedHistory, function(story) {
-        return _c("li", { key: story.year }, [
-          _c("span", [_vm._v(_vm._s(story.year))]),
-          _vm._v(" "),
-          _c(
-            "table",
-            _vm._l(story.historyList, function(list) {
-              return _c("tr", { key: list.month }, [
-                _c("td", [_vm._v(_vm._s(list.month))]),
-                _vm._v(" "),
-                _c("td", [_vm._v(_vm._s(list.description))])
-              ])
-            }),
-            0
-          )
-        ])
-      }),
-      0
-    ),
-    _vm._v(" "),
-    _c("button", { on: { click: _vm.isOpenHandler } }, [
-      _vm._v(
-        "\n        " + _vm._s(_vm.isOpen ? "접어두기" : "더보기") + "\n    "
-      )
-    ])
-  ])
-}
-var staticRenderFns = []
-render._withStripped = true
-
-
-
-/***/ }),
-
-/***/ "./resources/js/components/companyInfo/companyInfoComponents/History.vue":
-/*!*******************************************************************************!*\
-  !*** ./resources/js/components/companyInfo/companyInfoComponents/History.vue ***!
-  \*******************************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _History_vue_vue_type_template_id_7201f5a7___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./History.vue?vue&type=template&id=7201f5a7& */ "./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=template&id=7201f5a7&");
-/* harmony import */ var _History_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./History.vue?vue&type=script&lang=js& */ "./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=script&lang=js&");
-/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
-
-
-
-
-
-/* normalize component */
-
-var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
-  _History_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
-  _History_vue_vue_type_template_id_7201f5a7___WEBPACK_IMPORTED_MODULE_0__["render"],
-  _History_vue_vue_type_template_id_7201f5a7___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
-  false,
-  null,
-  null,
-  null
-  
-)
-
-/* hot reload */
-if (false) { var api; }
-component.options.__file = "resources/js/components/companyInfo/companyInfoComponents/History.vue"
-/* harmony default export */ __webpack_exports__["default"] = (component.exports);
-
-/***/ }),
-
-/***/ "./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=script&lang=js&":
-/*!********************************************************************************************************!*\
-  !*** ./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=script&lang=js& ***!
-  \********************************************************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_History_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../../node_modules/babel-loader/lib??ref--4-0!../../../../../node_modules/vue-loader/lib??vue-loader-options!./History.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=script&lang=js&");
-/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_History_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
-
-/***/ }),
-
-/***/ "./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=template&id=7201f5a7&":
-/*!**************************************************************************************************************!*\
-  !*** ./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=template&id=7201f5a7& ***!
-  \**************************************************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_History_vue_vue_type_template_id_7201f5a7___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../../node_modules/vue-loader/lib??vue-loader-options!./History.vue?vue&type=template&id=7201f5a7& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/companyInfo/companyInfoComponents/History.vue?vue&type=template&id=7201f5a7&");
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_History_vue_vue_type_template_id_7201f5a7___WEBPACK_IMPORTED_MODULE_0__["render"]; });
-
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_History_vue_vue_type_template_id_7201f5a7___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
-
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "companyInfoData", function() { return companyInfoData; });
+var companyInfoData = {
+  brandMission: [{
+    id: 1,
+    head: "여가 경험을 공유하는 <span class='point'>공간</span>",
+    description: "내가 기록한 여가 경험을 나에겐 추억이, 누군가에겐 영감이 되는 곳"
+  }, {
+    id: 2,
+    head: "혼자가 아닌 누군가와 함께 한다는 <span class='point'>즐거움</span>",
+    description: "내가 기록한 여가 경험을 나에겐 추억이, 누군가에겐 영감이 되는 곳"
+  }, {
+    id: 3,
+    head: "우리가 함께 만드는 <span class='point'>경험 지도</span>",
+    description: "우리의 경험을 모아 최대 규모의 공유 지도를 만드는 것"
+  }],
+  companyProfile: [{
+    head: "Company Name",
+    description: "ZziEut <span>찌읏</span>"
+  }, {
+    head: "CEO",
+    description: "Moon Seong-soo <span>문성수</span>"
+  }, {
+    head: "Main Services",
+    description: "Zzim, TsuRoop, ZzimStore <span>찜, 츄릅, 찜스토어</span>"
+  }, {
+    head: "Main Business",
+    description: "SNS Commerce Solutions, Fashion Design Platform, SI, Agency"
+  }],
+  contact: [{
+    head: "Address",
+    description: "8-12, Dongsomoonro 23-gil, Seongbuk-gu, Seoul, Republic of Korea (Jian Building, 4th floor)"
+  }, {
+    head: "Direction",
+    description: "7 minutes walk from exit 7 of Sungsin Women's University Station"
+  }, {
+    head: "Phone",
+    description: "070-4905-2942"
+  }, {
+    head: "Email",
+    description: "kkotkkio@zzieut.com"
+  }]
+};
 
 
 /***/ })
