+++ resources/sass/app.scss
@@ -12,3 +12,6 @@
 
 //CompanyInfo
 @import "./CompanyInfo";
+
+// @import "vue-slick-carousel/dist/vue-slick-carousel.css";
+// @import "vue-slick-carousel/dist/vue-slick-carousel-theme.css";
